refactor(codegen): extract polymorphic into plugin folder
